refactor: split ManageSubscriptionDialog
